Fix/perps optimize trades history (#8669)
* feat: optimize perps trades history with real-time WebSocket updates

- Replace HTTP polling with WebSocket subscription for trades history
- Add USER_FILLS subscription refresh mechanism for real-time updates
- Implement subscription handler pause/resume to optimize background resource usage
- Introduce EPerpsSubscriptionCategory enum for better type safety
- Refactor buildRequiredSubscriptionsMap to improve code maintainability
- Track new user fills when subscriptions are disabled and refresh on resume
- Optimize trades history merging logic to prevent duplicate data

This change significantly reduces server load and network latency while improving
real-time data accuracy for perps trading history.



* refactor: extract TradingView message handler to separate file

- Extract usePerpsMessageHandler to usePerpsTradingViewMessageHandler
- Create dedicated usePerpsTradingViewMessageHandler.ts file for better organization
- Add refreshWebviewMarksByApi method to handle marks refresh
- Update index.ts to use re-export pattern
- Add 'minmov' to spell checker skip words

This improves code modularity and makes the message handler logic easier to maintain
and test independently.



---------
