+++ packages/kit/src/views/Perp/hooks/usePerpOrderInfoPanel.ts
@@ -1,87 +1,90 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useRef, useState } from 'react';
 
-import { noop } from 'lodash';
+import { isEqual, noop } from 'lodash';
 
 import {
   usePerpsActiveAccountAtom,
   usePerpsTradesHistoryRefreshHookAtom,
 } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
-import {
-  PERPS_HISTORY_FILLS_URL,
-  PERPS_USER_FILLS_TIME_RANGE,
-} from '@onekeyhq/shared/src/consts/perp';
+import { PERPS_HISTORY_FILLS_URL } from '@onekeyhq/shared/src/consts/perp';
 import { appEventBus } from '@onekeyhq/shared/src/eventBus/appEventBus';
 import { EAppEventBusNames } from '@onekeyhq/shared/src/eventBus/appEventBusNames';
 import { openUrlInApp } from '@onekeyhq/shared/src/utils/openUrlUtils';
 import type { IFill, IWsUserFills } from '@onekeyhq/shared/types/hyperliquid';
-import { ESubscriptionType } from '@onekeyhq/shared/types/hyperliquid';
+import {
+  EPerpsSubscriptionCategory,
+  ESubscriptionType,
+} from '@onekeyhq/shared/types/hyperliquid';
 
 import backgroundApiProxy from '../../../background/instance/backgroundApiProxy';
-import { usePromiseResult } from '../../../hooks/usePromiseResult';
-
-interface INewTradesHistory {
-  fill: IFill;
-  userId: string | null;
-}
 
 export function usePerpTradesHistory() {
   const [currentAccount] = usePerpsActiveAccountAtom();
-  const [currentListPage, setCurrentListPage] = useState(1);
-  const [newTradesHistory, setNewTradesHistory] = useState<INewTradesHistory[]>(
-    [],
-  );
-  const [{ refreshHook }] = usePerpsTradesHistoryRefreshHookAtom();
-  const newTradesHistoryRef = useRef<INewTradesHistory[]>([]);
-  if (newTradesHistoryRef.current !== newTradesHistory) {
-    newTradesHistoryRef.current = newTradesHistory;
-  }
-  useEffect(() => {
-    if (
-      !currentAccount?.accountAddress ||
-      newTradesHistoryRef.current.length === 0
-    ) {
-      return;
-    }
-    const filterNewTradesHistory = newTradesHistoryRef.current.filter(
-      (trade) => trade.userId === currentAccount?.accountAddress,
-    );
-    setNewTradesHistory(filterNewTradesHistory);
-  }, [currentAccount?.accountAddress]);
-  useEffect(() => {
-    if (!currentAccount?.accountAddress) return;
+  const currentAccountAddressRef = useRef(currentAccount?.accountAddress);
+  currentAccountAddressRef.current = currentAccount?.accountAddress;
 
+  const [{ refreshHook }] = usePerpsTradesHistoryRefreshHookAtom();
+
+  const [isLoading, setIsLoading] = useState(true);
+  const [currentListPage, setCurrentListPage] = useState(1);
+  const [tradesHistory, setTradesHistory] = useState<IFill[]>([]);
+
+  const tradesHistoryRef = useRef<IFill[]>([]);
+  if (tradesHistoryRef.current !== tradesHistory) {
+    tradesHistoryRef.current = tradesHistory;
+  }
+
+  const tradesHistoryAccountRef = useRef<string | undefined>(undefined);
+
+  useEffect(() => {
     const handleUserFillsListUpdate = (payload: unknown) => {
       const eventPayload = payload as {
-        type: 'account';
-        subType: string;
+        type: EPerpsSubscriptionCategory;
+        subType: ESubscriptionType;
         data: IWsUserFills;
       };
 
-      if (eventPayload.subType !== ESubscriptionType.USER_FILLS) return;
-
       if (
-        !eventPayload?.data?.user ||
-        eventPayload?.data?.user?.toLowerCase() !==
-          currentAccount?.accountAddress?.toLowerCase()
+        eventPayload.type === EPerpsSubscriptionCategory.ACCOUNT &&
+        eventPayload.subType === ESubscriptionType.USER_FILLS
       ) {
-        return;
+        setIsLoading(false);
+        const { data } = eventPayload;
+        if (data.fills.length === 0) {
+          return;
+        }
+        const newFills = [...data.fills]
+          .filter((fill) => !fill.coin.startsWith('@'))
+          .sort((a, b) => b.time - a.time);
+        const newFirstFillItem = newFills[0];
+        const prevFirstFillItem = tradesHistoryRef.current[0];
+
+        const eventUserAddress = eventPayload.data?.user?.toLowerCase();
+        const currentAccountAddress =
+          currentAccountAddressRef.current?.toLowerCase();
+
+        if (eventUserAddress && eventUserAddress === currentAccountAddress) {
+          const isFirstItemEqual = isEqual(prevFirstFillItem, newFirstFillItem);
+          // check if first item is the same, then skip merge
+          if (isFirstItemEqual && prevFirstFillItem) {
+            return;
+          }
+          tradesHistoryAccountRef.current = currentAccountAddress;
+          if (data.isSnapshot) {
+            setTradesHistory((_prev) => [...newFills]);
+          } else {
+            setTradesHistory((prev) => [...newFills, ...prev]);
+          }
+        } else if (
+          prevFirstFillItem &&
+          tradesHistoryAccountRef.current &&
+          tradesHistoryAccountRef.current !== currentAccountAddress
+        ) {
+          // reset trades history if account not matched
+          tradesHistoryAccountRef.current = currentAccountAddress;
+          setTradesHistory([]);
+        }
       }
-
-      const { data } = eventPayload;
-
-      if (data.isSnapshot) return;
-
-      const relevantFills = [...data.fills];
-
-      if (relevantFills.length === 0) return;
-
-      setNewTradesHistory((prev) => [
-        ...prev,
-        ...relevantFills.map((fill) => ({
-          fill,
-          userId: currentAccount?.accountAddress,
-        })),
-      ]);
     };
 
     appEventBus.on(
@@ -95,54 +98,28 @@ export function usePerpTradesHistory() {
         handleUserFillsListUpdate,
       );
     };
-  }, [currentAccount?.accountAddress]);
-  const { result, isLoading } = usePromiseResult(
-    async () => {
-      if (currentAccount?.accountAddress) {
-        noop(refreshHook);
-        const now = Date.now();
-        const startTime = now - PERPS_USER_FILLS_TIME_RANGE;
-        const trades =
-          await backgroundApiProxy.serviceHyperliquid.getUserFillsByTime({
-            user: currentAccount?.accountAddress,
-            startTime,
-            aggregateByTime: true,
-          });
-        const sortedTrades = trades.sort((a, b) => b.time - a.time);
-        setCurrentListPage(1);
-        return sortedTrades;
-      }
-      return [];
-    },
-    [currentAccount?.accountAddress, refreshHook],
-    { watchLoading: true, initResult: [] },
-  );
+  }, []);
 
-  const mergeTradesHistory = useMemo(() => {
-    let mergedTrades = result;
-    if (newTradesHistory.length > 0) {
-      const existingTIds = new Set(result.map((trade) => trade.tid));
-      const newUniqueTrades = newTradesHistory
-        .filter(
-          (trade) =>
-            !existingTIds.has(trade.fill.tid) &&
-            trade.userId === currentAccount?.accountAddress,
-        )
-        .map((trade) => trade.fill);
-
-      if (newUniqueTrades.length === 0) {
-        return result;
-      }
-      mergedTrades = [...mergedTrades, ...newUniqueTrades];
+  useEffect(() => {
+    setCurrentListPage(1); // reset to page 1 when account changed
+    tradesHistoryAccountRef.current = '';
+    setTradesHistory([]);
+    if (currentAccount?.accountAddress) {
+      setIsLoading(true);
+    } else {
+      setIsLoading(false);
     }
-    const filteredTrades = mergedTrades.filter((t) => !t.coin.startsWith('@'));
-    return filteredTrades?.sort((a, b) => b.time - a.time);
-  }, [currentAccount?.accountAddress, newTradesHistory, result]);
+  }, [currentAccount?.accountAddress]);
 
-  console.log('usePerpTradesHistory__render', newTradesHistory, refreshHook);
+  useEffect(() => {
+    noop(refreshHook);
+    void backgroundApiProxy.serviceHyperliquidSubscription.updateSubscriptionForUserFills();
+  }, [refreshHook]);
+
+  console.log('usePerpTradesHistory__render', tradesHistory, refreshHook);
 
   return {
-    trades: mergeTradesHistory,
+    trades: tradesHistory,
     currentListPage,
     setCurrentListPage,
     isLoading,
